requirements.txt
3==0.0.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cPython==0.0.6
Django==4.2.3
django-admin-logs==1.0.2
django-nine==0.2.7
django-nonmodel-admin==0.1.1
django-strawberry==0.1.2
djangorestframework==3.14.0
dnspython==2.4.0
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
libcst==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.1
paynow==1.0.4
psycopg2-binary==2.9.6
Pygments==2.15.1
pymongo==4.4.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.4.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.30.0
strawberry-django-plus==1.11.2
strawberry-graphql==0.195.3
strawberry-graphql-django==0.9.5
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
